Move trigger mod to the separate file
